requirements.txt
asgiref==3.2.5
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
Django==3.0.4
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jwt==1.0.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.3
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
